feat(Snapshot): mf-6503 refine voting ui
