+++ packages/plugins/Snapshot/src/SiteAdaptor/ProgressTab.tsx
@@ -1,14 +1,18 @@
-import { ResultCard } from './ResultCard.js'
-import { VotesCard } from './VotesCard.js'
+import { MaskLightTheme } from '@masknet/theme'
+import { ThemeProvider } from '@mui/material'
 import { InformationCard } from './InformationCard.js'
+import { ResultCard } from './ResultCard.js'
 import { SnapshotTab } from './SnapshotTab.js'
+import { VotesCard } from './VotesCard.js'
 
 export function ProgressTab() {
     return (
         <SnapshotTab>
-            <InformationCard />
-            <ResultCard />
-            <VotesCard />
+            <ThemeProvider theme={MaskLightTheme}>
+                <InformationCard />
+                <ResultCard />
+                <VotesCard />
+            </ThemeProvider>
         </SnapshotTab>
     )
 }
